refactor: Move Vite virtual module plugin to dedicated file
